requirements.txt
flask
flask-cors
numpy
face-recognition
Pillow
requests
